Begin a large restructuring of models
